fix for active stream bookkeeping, decrement on used streams
the previous patch removed decrementing, which meant that only max
stream number of requests were possible on a connection; this is wrong,
as per the spec, streams which move to CLOSED state should not be
accounted as active; the original patch was motivated by solving an h2spec
compliance issue, given that upstream http-2 does not do it correctly
(it always decrements, including in the case where a push stream hasn't
transitioned to half-closed). This patch addresses both.

Closes #4.