bookkeep active streams count in connection straight from the stream
this has been historically done using the callback functions. This is
dangerous, because these are also exposed to the user of the lib. This
means that, when creating a stream, for instance a user can (and will)
had a handler for the "stream closed" event, which will run before the
internal handler doing bookkeeping. If the handler creates a new stream
it'll fail, because the bookkeeping hasn't happened yet, and "limit
exceeded" will be raised.

the previous patch made it worse by setting bookkeeping as a
conditional callback on the active callback. The user would now only be
able to set callbacks on "active" event as well. Not a great experience
IMO.

So this change will make it so that this happens outside of the callback
lifecycle, and the user does not need to wait on "active" to set
callbacks.